--- config/config.yaml
@@ -1,15 +1,0 @@
-scrape:
-  source_url: "https://xkcd.com/"
-  scrape_pages_limit: -1   # -1 -> all pages
-  request_retries: 3
-  parallel : 50
-
-database:
-  db_path : "./pkg/database/database.db"
-  index_path : "./pkg/database/index.json"
-  temp_dir : "./pkg/database/"
-  temp_folder_pattern : "temp_xkcd_"
-  temp_file_pattern : "response_xkcd" # do not use "-" symbol
-
-rest_api:
-  port: 222
